Refactor foph import module
