+++ epigraphhub/data/foph/loading.py
@@ -4,7 +4,6 @@ This module will retrieve the data from a CSV file, create a connection
 with the SQL Database and update it with the new information. Pangres
 will generate chunks with total length of 1000 and insert them into the
 corresponding table as specified by the downloaded CSV file.
-@see epigraphhub.data.data_collection.foph.download
 @see epigraphhub.connection
 
 Methods
@@ -13,10 +12,11 @@ Methods
 load(table, filename):
     Connects to SQL DB and update a table with CSV information.
 """
+from datetime import datetime
+
 import pandas as pd
 from loguru import logger
 from pangres import upsert
-from datetime import datetime
 
 from epigraphhub.connection import get_engine
 from epigraphhub.data._config import FOPH_CSV_PATH, FOPH_LOG_PATH
